+++ docs/html/search/all_2.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['getchords_0',['getChords',['../class_score.html#a4deac88e3d1f4d27e5fd18ac60840341',1,'Score']]]
+  ['getchords_0',['getChords',['../class_score.html#aa4d6fd98b8ed74790a79e1c8fb095689',1,'Score']]]
 ];